# swagger.yaml
openapi: 3.0.3
info:
  title: INTVoicekit API
  description: |
    Backend API for INTVoicekit - a voice assistant platform built on Express.js with MongoDB.
    This API acts as a proxy layer between clients and the external Vyom LiveKit API,
    managing local state for users, assistants, SIP trunks, tools, inbound configurations,
    and analytics.

    ## Authentication
    Every endpoint except `POST /api/auth/signup` and `POST /api/auth/login` requires the user's
    upstream LiveKit key, sent as `Authorization: Bearer <api_key>`. A missing or unknown key is
    a `401`. Obtain the key from signup or login. Identity comes from that key — a `user_id` in
    the request body or query string is ignored, and the key is forwarded to the external Vyom
    API on the user's behalf.

    ## Dual ID Resolution
    Many endpoints accept either a local MongoDB `_id` or an external service ID
    (e.g., `external_assistant_id`, `external_trunk_id`).

    ## External API Proxy
    Most service methods proxy requests to `https://api-livekit-vyom.indusnettechnologies.com/*`,
    forwarding the user's Bearer token. External responses are typically passed through directly.

    ## Assistant Runtime Modes
    `assistant_mode` selects the session shape. It is `pipeline` by default.
    - `pipeline` (half-cascade): a realtime model emits text, an external TTS provider speaks it.
      Vendor is `openai` only — `gemini` is rejected with `400` (see below). Requires
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/assistant/create:
    post:
      tags: [Assistant]
      summary: Create a new voice assistant
      description: |
        Creates a voice assistant both locally in MongoDB and externally via the Vyom API.
        Supports runtime modes `pipeline`, `realtime`, and `cascade`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assistant_name, assistant_description, assistant_prompt]
              properties:
                assistant_name:
                  type: string
                  example: "Customer Support Bot"
                assistant_description:
                  type: string
                  example: "Handles inbound customer support queries"
                assistant_prompt:
                  type: string
                  example: "You are a helpful customer support agent..."
                assistant_mode:
                  type: string
                  enum: [pipeline, realtime, cascade]
                  default: pipeline
                  example: "pipeline"
                assistant_llm_config:
                  allOf:
                    - $ref: '#/components/schemas/AssistantLlmConfig'
                    - description: Required when mode=realtime, optional in pipeline.
                assistant_tts_model:
                  type: string
                  description: Required in pipeline mode. Must be sent together with assistant_tts_config.
                  enum: [sarvam, cartesia, elevenlabs, mistral]
                  example: "sarvam"
                assistant_tts_config:
                  $ref: '#/components/schemas/AssistantTtsConfig'
                assistant_stt_model:
                  type: string
                  description: >
                    User-transcription source. `native` (the realtime LLM transcribes itself)
                    is pipeline-only and rejected in cascade. The four plugin providers
                    (`cartesia`, `deepgram`, `elevenlabs`, `openai`) run for real in cascade;
                    in pipeline they are stored but that call falls back to native
                    transcription. `sarvam` runs in both. `realtime` ignores STT fields
                    because the model transcribes directly.
                  enum: [sarvam, native, cartesia, deepgram, elevenlabs, openai]
                  default: sarvam
                assistant_stt_config:
                  $ref: '#/components/schemas/AssistantSttConfig'
                assistant_start_instruction:
                  type: string
                  example: "Greet the user warmly"
                assistant_interaction_config:
                  $ref: '#/components/schemas/AssistantInteractionConfig'
                assistant_end_call_enabled:
                  type: boolean
                  example: false
                assistant_end_call_trigger_phrase:
                  type: string
                  example: "goodbye"
                assistant_end_call_agent_message:
                  type: string
                  example: "Thank you for calling. Goodbye!"
                assistant_end_call_url:
                  type: string
                  format: uri
                  example: "https://example.com/webhook/end-call"
                assistant_end_call_webhook:
                  $ref: '#/components/schemas/AssistantEndCallWebhook'
                assistant_greeting_audio:
                  type: object
                  properties:
                    enabled:
                      type: boolean
                      example: true
                    audio_id:
                      type: string
                      example: "audio_123"
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '201':
          description: Assistant created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Assistant created successfully"
                  assistant:
                    $ref: '#/components/schemas/Assistant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/list:
    get:
      tags: [Assistant]
      summary: List all assistants
      description: Retrieves a paginated list of assistants from the external API.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 100
          description: >
            Page size. Defaults to 100 here — deliberately higher than the external API's
            own default of 10, because existing clients expect the full list in one call.
        - name: assistant_name
          in: query
          schema:
            type: string
          description: Case-insensitive partial name filter.
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
          description: ISO-8601 lower bound on assistant creation time.
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
          description: ISO-8601 upper bound on assistant creation time.
        - name: sort_by
          in: query
          schema:
            type: string
            default: assistant_created_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Assistant list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Assistant'
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/details/{id}:
    get:
      tags: [Assistant]
      summary: Get assistant details
      description: Retrieves details for a specific assistant by local or external ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Assistant details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assistant'
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Assistant not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/update/{id}:
    patch:
      tags: [Assistant]
      summary: Update an assistant
      description: Updates an assistant locally and on the external Vyom API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assistant_name:
                  type: string
                assistant_description:
                  type: string
                assistant_prompt:
                  type: string
                assistant_mode:
                  type: string
                  enum: [pipeline, realtime, cascade]
                  description: >
                    Target mode. Only this field (or the presence of TTS/STT fields) switches
                    mode — sending assistant_llm_config alone does NOT, and leaves the stored
                    TTS/STT config untouched.
                assistant_llm_config:
                  allOf:
                    - $ref: '#/components/schemas/AssistantLlmConfig'
                    - description: >
                        Required only when switching into realtime. Safe to send on its own in
                        pipeline mode to rotate api_key or pick gemini.
                assistant_tts_model:
                  type: string
                  enum: [sarvam, cartesia, elevenlabs, mistral]
                  description: >
                    Only required when no TTS config is stored yet. Either half of the TTS pair
                    is enough — the other is filled in from the stored assistant, so changing
                    just the config keeps the model and vice versa. Switching provider without
                    a new config resets to that provider's defaults.
                assistant_tts_config:
                  $ref: '#/components/schemas/AssistantTtsConfig'
                assistant_stt_model:
                  type: string
                  enum: [sarvam, native, cartesia, deepgram, elevenlabs, openai]
                  description: >
                    `native` is pipeline-only and rejected in cascade; `sarvam` is valid in
                    both; `cartesia`, `deepgram`, `elevenlabs` and `openai` run in cascade and
                    are stored-but-inert in pipeline. Sending the model alone resets that
                    provider's config to defaults. Switching an assistant to cascade while it
                    still holds `native` is a 400 — send the new STT model in the same request.
                assistant_stt_config:
                  $ref: '#/components/schemas/AssistantSttConfig'
                assistant_start_instruction:
                  type: string
                assistant_interaction_config:
                  $ref: '#/components/schemas/AssistantInteractionConfig'
                assistant_end_call_enabled:
                  type: boolean
                assistant_end_call_trigger_phrase:
                  type: string
                assistant_end_call_agent_message:
                  type: string
                assistant_end_call_url:
                  type: string
                assistant_end_call_webhook:
                  $ref: '#/components/schemas/AssistantEndCallWebhook'
                assistant_greeting_audio:
                  type: object
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Assistant updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Assistant updated successfully"
                  data:
                    type: object
                    properties:
                      assistant_id:
                        type: string
                  local_data:
                    $ref: '#/components/schemas/Assistant'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/delete/{id}:
    delete:
      tags: [Assistant]
      summary: Delete an assistant
      description: Deletes an assistant from both local DB and external API.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Assistant deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Assistant deleted successfully"
                  data:
                    type: object
                    properties:
                      assistant_id:
                        type: string
                  local_data_removed:
                    type: boolean
                    example: true
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/call-logs/{id}:
    get:
      tags: [Assistant]
      summary: Get assistant call logs
      description: Retrieves paginated call logs for a specific assistant.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Call logs retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/billable-minutes/{id}:
    get:
      tags: [Assistant]
      summary: Get billable minutes for an assistant
      description: Calculates total billable minutes for a specific assistant and phone number.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: to_number
          in: query
          required: true
          schema:
            type: string
          example: "+1234567890"
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Billable minutes calculated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Total billable minutes calculated successfully"
                  data:
                    type: object
                    properties:
                      assistant_id:
                        type: string
                      to_number:
                        type: string
                      total_billable_minutes:
                        type: number
                        example: 125.5
                        description: >
                          Renamed from total_billable_duration_minutes so this endpoint and
                          platform-billable-minutes use one key for the same quantity.
                      timespan_evaluated:
                        type: object
                        properties:
                          start_date:
                            type: string
                          end_date:
                            type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/platform-billable-minutes:
    get:
      tags: [Assistant]
      summary: Get platform-wise billable minutes
      description: >-
        Calculates billable minutes grouped by platform phone number.

        There is no upstream aggregate for this — it is computed here by reading every call log for
        every assistant you own, so it is the slowest endpoint in the API and its cost grows with
        your history. Always send `start_date` and `end_date`; omitting them evaluates your whole
        lifetime.

        A per-assistant read failure is skipped rather than failing the request, so check
        `data.assistants_skipped` before treating the totals as complete — a non-zero value means
        the numbers are lower than reality, and `message` says so too.
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Platform billable minutes retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Platform-wise billable minutes calculated successfully"
                  data:
                    type: object
                    properties:
                      platform_wise_minutes:
                        type: array
                        items:
                          type: object
                          properties:
                            platform_number:
                              type: string
                            total_billable_minutes:
                              type: number
                      timespan_evaluated:
                        type: object
                        properties:
                          start_date:
                            type: string
                          end_date:
                            type: string
                      assistants_evaluated:
                        type: integer
                        description: How many assistants were included in this total.
                        example: 12
                      assistants_skipped:
                        type: integer
                        description: >-
                          How many assistants could not be read and were left out. Anything above 0
                          means the totals are incomplete.
                        example: 0
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/assistant/platform-billable-minutes/download:
    get:
      tags: [Assistant]
      summary: Download platform billable minutes as Excel
      description: Downloads an Excel file (.xlsx) with platform-wise billable minutes.
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Excel file download
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== SIP ====================
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Something went wrong"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The user's upstream LiveKit `api_key`, obtained from `POST /api/auth/signup` or
        `POST /api/auth/login`, sent as `Authorization: Bearer <api_key>`.

# Applied to every operation; /api/auth/signup and /api/auth/login override it with `security: []`.
security:
  - bearerAuth: []